Update a bunch of tests
